+++ ChangeLog.md
@@ -1,6 +1,6 @@
 # CHANGELOG REMOVECONF FOR <a href="https://www.dolibarr.org">DOLIBARR ERP CRM</a>
 ## 2.1.0
-New suppress BOM confirmation pop-ups</br>
+New suppress BOM confirmation pop-ups
 New suppress MO confirmation pop-ups
 
 ## 2.0.2